Move pkg.File to snapfile.File
